% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.LiangSeg-package.R
\docType{package}
\name{addin.LiangSeg-package}
\alias{addin.LiangSeg}
\alias{addin.LiangSeg-package}
\title{addin.LiangSeg: Addin to the package rABM for Hao Liang's Segregation by Landlords model}
\description{
This addin requires KeiichiSatoh/rABM from github to run.
}
\author{
\strong{Maintainer}: Keiichi Satoh \email{keiichi.satoh@r.hit-u.ac.jp}

}
\keyword{internal}
